Updated big integer library

+++ include/alef/lib/concepts/numeric.h
@@ -3,7 +3,7 @@
 
 #include "__detail.h"
 
-namespace alf::concepts::numeric {
+namespace alf::concepts {
 
 template<typename T>
 concept integral = alf::concepts::__detail::is_integral_v<T>;
@@ -20,6 +20,6 @@ concept floating_point = alf::concepts::__detail::is_floating_point_v<T>;
 template<typename T>
 concept number = floating_point<T> || integral<T>;
 
-} // namespace alf::concepts::numeric
+} // namespace alf::concepts
 
 #endif // ALEF_CORE_CONCEPTS_NUMERIC_H
